# Conflicts:
#	version

+++ src/main/java/it/gov/pagopa/payhub/activities/activity/classifications/ClearClassifyIufActivity.java
@@ -16,5 +16,5 @@ public interface ClearClassifyIufActivity {
      * @return boolean true if success deletion or exception
      */
     @ActivityMethod
-    Long deleteClassificationByIuf(Long organizationId, String iuf);
+    Integer deleteClassificationByIuf(Long organizationId, String iuf);
 }
